drm/i915: Don't register CRT connectro when DDI E can't be used
On HSW/BDW DDI A and E share 2 lanes, so when DDI A requires the
shared lanes DDI E can't be used. The lanes are not suppsoed to
be dynamically switched between the two uses, so there's no point
in registering the CRT connector when DDI E has no lanes.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>