drm/i915: Use fb->pitches[0] in cursor code
The cursor code currently ignores fb->pitches[0] (except when creating
the fb itself), and just uses the cursor_width*4 as the stride. Let's
make fb->pitches[0] actually matches what we expect it to.

We can also relax the stride vs. cursor width relationship on 845/865
since the stride is programmed separately. The only constraint is that
'cursor width*cpp <= stride' and that's already taken care of by
earlier checks that make sure 'fb->width*cpp <= stride' and
'src_w <= fb->width'.

We can also drop the bo size check as that's already checked when
we create the fb.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>