drm/i915: Enable vblank_disable_immediate on gen2
Since
commit 4dfd648 ("drm: Use vblank timestamps to guesstimate how many vblanks were missed")
the vblank code can cook up a frame counter value based on
the vblank timestamps (as long as they're accurate), so there's
no longer any need to keep vblank interrupts enabled on gen2
when no one is interested in them. So let's opt into the
immediate disable scheme on gen2 as well.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>