drm/i915: Register shadow VGA even when it produces spurious detection results
Having a shadow VGA connector is useful for testing purposes. We
curently skip registering the connector on machines where the
CRT detect falsely reports it as connected. Let's instead move the
the blacklist check to the detect callback (and hpd setup) and
if we get a match we always report the connector as disconnected.
This way we get a shadow VGA connector to help with testing, while
we still avoid the user facing problems from the incorrect
detection results.

commit 8ca4013d702d ("CHROMIUM: i915: Add DMI override to skip CRT initialization on ZGB")
doesn't provide much in the way of details as to why 'ACER ZGB' was
added to the blacklist. Trying to trace it further leads me to a
chromeos bugreport I can't access. So based on the fact that the commit added
the "/* Skip machines without VGA that falsely report hotplug events */"
comment, I'm going to assume that it was just spurious CRT detection. So
it should be safe to move the blacklist to just block the detection and hpd
without causing a regression on said machine.

Cc: Duncan Laurie <dlaurie@chromium.org>
Cc: Olof Johansson <olofj@chromium.org>
Cc: Stéphane Marchesin <marcheu@chromium.org>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>