drm/i915: Nuke HAS_PIPE_CXSR
The automagic big-FIFO refresh rate switching was only available on
ctg/elk. Based on some notes in the ilk/snb docs it's no longer
supported, although most of the relevant bits are still documented.
There are references to some rate selection register that's nowhere
to be found however.

Let's assume the automagic mode only ever worked on ctg/elk, and
since we can use the manual mode even on those platforms, it seems
better to not use the automagic mode at all. Let's rip it out.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>