drm/i915: Add DPLL based refresh rate switching
A third method to change the refresh rate is via the DPLL FPx0/1
divider sets. This method is supported on all pre-ilk platforms.

FIXME: figure put the bw/cl situaltion. The spec says the bit
is reserved. It's not reserved on gen3 or ctg/elk, so bw/cl being
excluded would seem quite weird.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>