dim: add cherry-pick-[next-]fixes subcommands
Add two new subcommands for cherry-picking fixes from dinq to
drm-intel-fixes and drm-intel-next-fixes. The only difference in the
subcommands is the assert branch check to ensure the user is on the
right branch.

The commands scan dinq for commits Cc'd to stable or drm-intel-fixes,
checks whether they've already been backported, attempts cherry-pick,
and asks for directions on failed cherry-pick.

It's still rough around the edges and slow as molasses due to unlimited
scan for backports (should just check a release or two back at most),
and an aborted run just starts over next time (though it should take the
freshly backported commits into account).

Signed-off-by: Jani Nikula <jani.nikula@intel.com>