drm/i915: Keep track of last desired vco frequency
Now that skl_vco_freq tracks the actual DPLL0 vco frequency, we'll need
something that keeps track of which vco frequency we want to use in case
the current vco is 0. This would be important across supend/resume since
we'll disable DPLL0 around those parts. I suppose later we might want to
consider disabling DPLL0 also while no pipes are active, though the DMC
perhaps makes that semi-pointless as at least some versions of the
firmware for the DPLL0 back on even if it was disabled when the state
was saved.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>