dim: update for-next branches as part of nightly rebuild
Currently the for-linux-next-fixes and for-linux-next branches only get
updated after the nightly rebuild after pushing branches. If the nightly
rebuild fails at that point, the for-next branches won't get updated
until the next time something is pushed, not upon nightly rebuild. Fix
this by making for-next branch update part of dim rebuild-nightly.

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>