dim: run checkpatch on committed patches on apply
The input messages may have base64 encoding and whatnot, and
checkpatch.pl can't cope with them. Just let 'git am' handle that. The
upside is that checkpatch will now catch e.g. duplicate signed-off-bys.

This is a partial revert of

commit a913db697bb063d374229e5e806e514fa44985d0
Author: Daniel Vetter <daniel.vetter@ffwll.ch>
Date:   Wed Oct 14 18:32:49 2015 +0200

    dim: Store patch in a temp file in apply-patch

although we'll still keep the temp file for other purposes, such as
extracting the message-id.

While at it, drop the obsolete check for drm_i915_private_t. We'll now
get a build fail for that.

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>