drm/i915: Account for sink max TMDS clock when checking the port clock
Account for the max TMDS clock declared in the HDMI VSDB when
determining the max useable port clock.

It's perfectly legal for the sink to eg. support only 8bpc with
high resolution modes and 12bpc only with low resolution modes.
In that case the 12bpc port clock with the high resolution modes
could exceed the max TMDS clock, so we should not attemot to use
12bpc with these modes.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>