drm/i915: Track HDMI/infoframes/audio on a per-port basis
g4x can't send infoframes or audio to more than one port at a time.
To that end start to globally track which port is using those
features. We hand the infoframe/audio capability to ports on a first
come, first serve basis.

TODO: Not sure this actually works since I rebased it
      Moar cleanup needed

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>