drm/i915: Try to get EDID via ACPI _DDC if i2c fails for LVDS/eDP
ACPI has the _DDC method which can be used to get an EDID for a panel
if the panel doesn't actually have an EDID itself. Try to use the ACPI
way if the i2c EDID read fails.

Somewhat stolen from nouveau, but for extra paranoia I went and used
drm_do_get_edid() to make sure we validate the extension block count and
whatnot.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>