drm/i915: Clear pipestat consistently
We have a lot of different ways of clearing the PIPESTAT registers.
Let's unify it all into one function. There's no magic in PIPESTAT
that would require any of the double clearing and whatnot that
some of the code tries to do. All we can really do is clear the status
bits and disable the enable bits. There is no way to unmask anything
so as soon as another even happens the status bit will become 1 again,
and this trying to clear them twice or something won't help at all.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>