drm/i915: Do CRT HPD force trigger only when needed
There doesn't seem to be any need to for the force trigger trick every
single time on VLV CRT output. Let's do it the same way as on ILK+, and
do the force trigger only on the first detect after reset.

This means we can actually just call the ILK+ code on VLV. The only
difference between the two is the "turn DAC off" trick, but that's
already conditional on HAS_PCH_SPLIT() so it won't be done on VLV.

Cc: Lyude <cpaul@redhat.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1460995405-6654-1-git-send-email-ville.syrjala@linux.intel.com
Cc: stable@vger.kernel.org
Reviewed-by: Lyude <cpaul@redhat.com>