drm/i915: Always power on TX wells before CMN well on VLV
Supposedly we need to power on the TX wells before CMN well on VLV.
We currently do that when enabling HDMI/DP ports, but when we enable
the CRT port we only power up the CMN well. So if someone enables the
CRT port first, and then HDMI/DP the wells will power up in the supposed
wrong order.

I can't actually reproduce any issues when I power on the wells in the
wrong order, so I'm not 100% convinced we couldn't just go back to the
original order. But I suppose it's possible there are machines that are
more sensitive, so let's make sure we follow the order we know to work.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1460995405-6654-3-git-send-email-ville.syrjala@linux.intel.com
Cc: stable@vger.kernel.org
Reviewed-by: Lyude <cpaul@redhat.com>