drm/i915: Ask the sink whether training is required when exiting PSR main-link off mode
The sink can tell us if link training needs to be performed when
exiting PSR main-link off mode. Currently we get that information
from the VBT, but at least on my HSW the VBT says one thing, the sink
another. And in practice the sink doesn't seem to notice any screen
updates unless we do the training.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>