drm/i915: Ignore initial lid state for eDP
Currently we try to check the lid status to see whether eDP is connected
or not. However we ignore the result anyway, and we have no lid notifier
to fire off uevents, so let's just ignore the lid stuff for eDP
entirely.

While at it, rip out the eDP case from intel_dp_detect_dpcd() since we
never call that thing with eDP displays anyway.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>