drm/i915: Make primary/sprite zpos configurable on VLV/CHV
VLV/CHV can change the z order between the primary and sprite planes
freely. Let's expose that capability by making the zpos property
mutable. The cursor plane is always on top, so the cursor zpos is
left as an immutable property.

The way the hardware is configured is a bit awkward. Both sprite planes
have a "bottom" and "zorder" bits, but the way they interact betweens
the sprites doesn't seem entirely logical. The spec doesn't even list
all 16 possible combinations (just 10 are listed), but I've gone through
them all the discovered the following relationship:

sprite 0	sprite 1	resulting plane order (bottom to top)
B Z		B Z
0 0		0 0		P 0 1	  d0 d1     d01
0 0		0 1		P 0 1	*
0 0		1 0		1 P 0	* d0
0 0		1 1		1 P 0
0 1		0 0		P 1 0	*
0 1		0 1		P 0 1
0 1		1 0		1 0 P	*       dp
0 1		1 1		1 0 P
1 0		0 0		0 P 1	*    d1
1 0		0 1		0 1 P	*       dp
1 0		1 0		P 0 1
1 0		1 1		P 0 1
1 1		0 0		0 P 1
1 1		0 1		0 1 P
1 1		1 0		P 0 1
1 1		1 1		P 0 1

[B=bottom bit Z=zorder bit, 0=sprite 0, 1=sprite 1, P=primary,
 d0=sprite 0 disabled, d1=sprite 1 disabled, dp=primary disabled,
 d01=both sprites (and primary potentially) disabled]

Disabling any one or two planes doesn't seem to interfere with the
resulting plane order in an unexpected way. The bottom and zorder
bits of the disabled planes are still effective though, so we must be
careful what we leave in those bits when the plane is disabled.

What we do for enabled sprite planes is set bottom=1 if the plane
at the bottom, else we set zorder=1 if the plane is just above
the other sprite plane. This will end up picking the configurations
marked with * from the table, assuming both sprite planes are enabled.

For the disabled plane cases we just leave bottom=0 and zorder=0
for any disabled sprite lane. Any enabled sprite plane will still
follow the rule laid out before. For the purposes of that rule,
we'll consider any disabled plane to be stacked on top of the
enabled planes. This will end up picking the configurations marked
d0 when sprite 0 is disabled, d1 when sprite 1 is disabled,
dp when the primary is disabled, and d01 when both sprites are
disabled (and potentially the primary as well).

The use of the 'P 0 1' configuration for the d0 case doesn't really
make sense based on the above, but given the way the zorder bits work,
there's no nice way to pick the 'P 1 0' (which would be the logical
thing) when sprite 0 is disabled as that would require setting the
zorder bit in the sprite 0 control register, but as stated we always
set both bits to 0 for disabled planes. That way we don't have to
reconfigure already disabled planes to change the zorder between the
currently active planes. This slight irregularity doesn't matter in
the end as both of these configuratios will gives us the expected
visible 'P 1' result.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>