drm/i915: Move DP link retraining into intel_dp_detect()
DP link retraining needs to grab some modeset locks to not race with
modesets, so we can't really do it safely from the hpd_pulse, lest we
risk deadlocking due to MST sideband stuff.

Move the link retraining to happen from the hotplug work instead.
Doing at the end of intel_dp_detect() seems like a good place in case
the sink already got disconnected, in which case retraining is
pointless.

To determine if we need to schedule the hotplug work, we'll just check
the sink lane status without locks from hpd_pulse. A little racy
still eg. due to useing intel_dp->lane_count, but no less racy than
what we already had. We'll repeat the check in from intel_dp_detect()
with proper locking, where we'll also check if the link as actually
active or not.

Cc: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>
Cc: Jim Bride <jim.bride@linux.intel.com>
Cc: Manasi D Navare <manasi.d.navare@intel.com>
Cc: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>