drm/i915: Shut down displays gracefully on reboot
Dell UP2414Q doesn't like it when we're driving it in MST mode, and then
reboot the machine. After reboot, the monitor is somehow confused and
refuses to do the payload allocation:
 [drm:drm_dp_mst_allocate_vcpi] initing vcpi for 282 8
 [drm:drm_dp_dpcd_write_payload] status not set after read payload table status 0
and thus we're left with a black screen until the monitor power is
toggled.

It seems that if we shut down things gracefully prior to rebooting, the
monitor doesn't get confused like this. So let's try to shut down all
the displays gracefully on reboot. The downside is that we will
introduce the reboot notifier to all the modesetl locks. So if there's
a locking bug around, we might not be able to reboot the machine
gracefully. sysrq reboot will still work though, as it will not
call the notifiers.

While we do this, we can eliminate the eDP reboot notifier, which is
there to shut down the panel power and make sure the firmware won't
violate the panel power cycle delay post-reboot. Since we're now
shutting eDP down properly, we can mostly just rip out the eDP notifier.
We still need to do the panel power cycle wait though, as that would
normally get postponed until the next modeset. So let's move that part
into intel_panel so that other display types will get the same treatment
as a bonus.

The Dell UP2414Q can often get even more confused, and sometimes what
you have to do is: switch to another input on the monitor, toggle the
monitor power, switch the input back to the original setting. And
sometimes it seems you just have to yank the power cable entirely. I'm
not sure if this reboot notifier might avoid some of these other
failure modes as well, but I'm pretty sure it can't hurt at least.

v2: Add power cycle delay for LVDS not that we have the PPS delays
    Add a FIXME about needing a more thorough suspend to avoid VDD
    turning on due to AUX/HPD/etc.

Cc: Clint Taylor <clinton.a.taylor@intel.com>
Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>