drm/i915: Add memory barriers for SWSCI mailbox handling
We must make sure the SWSCI mailbox hsa the right contents when we
trigger the SWSCI. Add a write barriers to make that happen. One may
hope that the BIOS side has a corresponding read barrier.

Also we need a read barrier between reading the SCIC and PARM after
the BIOS has completed the operation.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>