dim: Stop recursive rr-cache from spreading
I had a bug in dim for short while which resulted in the rr-cache dir
getting copied into itself recursively, each time dim rebuild-nightly
was run. I fixed that, but unfortunately it spread to the real repo
already, and a bunch of folks have it in their local checkout.

Stop this infection from spreading by explicitly deleting the
recursive rr-cache dir.

Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>