drm/edid: Don't send bogus aspect ratios in AVI infoframes
If the user mode would specify an aspect ratio other than 4:3 or 16:9
we would end up stuffing that straight into the AVI infoframe as is,
assuming we found a VIC with a matching aspect ratio. That's not such
a good idea when the infoframe can't actually carry anything but
4:3 or 16:9. So let's not do that, and instead we'll leave the infoframe
aspect ratio unset in that case. The display should still figure it
out based on the VIC.

Not quite sure what would be the best way to handle the case when
the user has specified an aspect ratio that can't be conveyed via
the infoframe and we can't find a matching VIC. Here I've opted to
simply return an error. Not sure such a drastic approach is warranted
or if we should just silently ignore the user's wishes?

This isn't actually a problem so far as we don't have any modes that
would use one of the new aspect ratios in our cea mode list. But once
those are added this will become an issue.

Also toss in a warn to the infoframe packing code to let the developer
know they're trying to pass in an unsupported aspect ratio.

Cc: Shashank Sharma <shashank.sharma@intel.com>
Cc: Sean Paul <seanpaul@chromium.org>
Cc: Jose Abreu <Jose.Abreu@synopsys.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Emil Velikov <emil.l.velikov@gmail.com>
Cc: Thierry Reding <thierry.reding@avionic-design.de>
Fixes: a68362fe3e84 ("drm: Add and handle new aspect ratios in DRM layer")
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>